backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
asyncpg==0.31.0
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
requests==2.32.5
supabase==2.24.0
pydantic==2.12.5
